+++ audionimbus/CHANGELOG.md
@@ -84,7 +84,7 @@
 - Implement `Copy`, `Clone` for `SimulationError`.
 - Implement `Copy` for `ParameterValidationError`.
 - Implement `Clone` for `Spatialization`, `PathEffectSettings`, `VirtualSurroundEffectSettings`, `VirtualSurroundEffectParams`, `AmbisonicsBinauralEffectSettings`, `AmbisonicsBinauralEffectParams`, `AmbisonicsDecodeEffectSettings`, `AmbisonicsDecodeEffectParams` and `ReconstructorInputs`.
-- Implement `PartialEq`, `Eq` and `Hash` for `Scene` and `Simulator`.
+- Implement `PartialEq`, `Eq` and `Hash` for `Scene`, `Simulator`, `Context`, `EmbreeDevice`, `OpenClDevice`, `OpenClDeviceList`, `RadeonRaysDevice`, `TrueAudioNextDevice`, `ImpulseResponse`, `SerializedObject`, `PanningEffect`, `PathEffect`, `BinauralEffect`, `ReflectionEffect`, `ReflectionMixer`, `VirtualSurroundEffect`, `DirectEffect`, `AmbisonicsPanningEffect`, `AmbisonicsBinauralEffect`, `AmbisonicsEncodeEffect`, `AmbisonicsDecodeEffect`, `AmbisonicsRotationEffect`, `Hrtf`, `EnergyField`, `ProbeArray`, `ProbeBatch`, `Reconstructor`, `InstancedMesh`, `StaticMesh`.
 
 ### Removed